+++ web/wiki/src/openapi/openapiFetcher.ts
@@ -1,15 +1,35 @@
 import qs from 'qs'
-import {
-  type ListResourcesQueryParams,
-  type ListTopicsQueryParams,
-} from './openapiComponents'
 import { OpenapiContext } from './openapiContext'
 
 const baseUrl = '' // TODO add your baseUrl
 
 export type ErrorWrapper<TError> =
   | TError
-  | { status: 'unknown'; payload: string }
+  | { status: number | string; payload: string }
+
+class OpenapiError<TError> extends Error {
+  constructor(public error: ErrorWrapper<TError>) {
+    super()
+  }
+}
+
+async function processError<TError>(
+  response: Response
+): Promise<ErrorWrapper<TError>> {
+  let error: ErrorWrapper<TError>
+  try {
+    error = {
+      status: response.status,
+      payload: await response.json(),
+    }
+  } catch (e) {
+    error = {
+      status: 'unknown' as const,
+      payload: 'Unexpected error',
+    }
+  }
+  return error
+}
 
 export type OpenapiFetcherOptions<TBody, THeaders, TQueryParams, TPathParams> =
   {
@@ -79,7 +99,7 @@ export async function openapiFetch<
     if (!response.ok) {
       let error: ErrorWrapper<TError>
       try {
-        error = await response.json()
+        error = await processError<TError>(response)
       } catch (e) {
         error = {
           status: 'unknown' as const,
@@ -89,8 +109,7 @@ export async function openapiFetch<
               : 'Unexpected error',
         }
       }
-
-      throw error
+      throw new OpenapiError(error)
     }
 
     if (response.headers.get('content-type')?.includes('json')) {
@@ -100,37 +119,27 @@ export async function openapiFetch<
       return (await response.blob()) as unknown as TData
     }
   } catch (e) {
-    let errorObject: Error = {
-      name: 'unknown' as const,
-      message:
-        e instanceof Error ? `Network error (${e.message})` : 'Network error',
-      stack: e as string,
+    if (e instanceof OpenapiError) {
+      throw e.error
+    } else {
+      let errorObject: Error = {
+        name: 'unknown' as const,
+        message:
+          e instanceof Error ? `Network error (${e.message})` : 'Network error',
+        stack: e as string,
+      }
+      throw errorObject
     }
-    throw errorObject
   }
 }
 
-const buildQueryStringResources = ({
-  categorySlug,
-  topicSlug,
-  resourceTypes,
-  topic,
-  status,
-  search,
-}: ListResourcesQueryParams) =>
-  qs.stringify({
-    categorySlug,
-    topicSlug,
-    resourceTypes,
-    topic,
-    status,
-    search,
-  })
-
-const buildQueryStringTopics = ({ categoryId, slug }: ListTopicsQueryParams) =>
-  qs.stringify({
-    categoryId,
-    slug,
+const buildQueryString = (queryParams: Record<string, string>) =>
+  qs.stringify(queryParams, {
+    skipNulls: true,
+    filter: (_prefix, value) => {
+      if (value === '') return undefined
+      return value
+    },
   })
 
 const resolveUrl = (
@@ -139,13 +148,6 @@ const resolveUrl = (
   pathParams: Record<string, string> = {}
 ) => {
   let query = new URLSearchParams(queryParams).toString()
-
-  if (queryParams as ListResourcesQueryParams) {
-    query = buildQueryStringResources(queryParams)
-  } else if (queryParams as ListTopicsQueryParams) {
-    query = buildQueryStringTopics(queryParams)
-  }
-
   if (query) query = `?${query}`
   return url.replace(/\{\w*\}/g, (key) => pathParams[key.slice(1, -1)]) + query
 }