+++ services/wiki/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.18.3] - 2024-07-24
+
+### Fixed
+
+– Resolved Swagger path parameters for `resources/favorites/{categorySlug}`
+
 ## [0.18.2] - 2024-07-23
 
 ### Added